+++ src/lib.rs
@@ -52,11 +52,13 @@ This crate also contains two "in development" structs — [`NicePercent`] and [`
 
 
 
+mod hash;
 pub(crate) mod macros;
 mod nice_elapsed;
 mod nice_int;
 pub mod traits;
 
+pub use hash::NoHash;
 pub use nice_elapsed::NiceElapsed;
 pub use nice_int::{
 	nice_u8::NiceU8,