+++ src/traits/mod.rs
@@ -2,8 +2,10 @@
 # Dactyl: Traits
 */
 
+mod btoi;
 mod btou;
 mod saturating_from;
 
+pub use btoi::BytesToSigned;
 pub use btou::BytesToUnsigned;
 pub use saturating_from::SaturatingFrom;