+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 
 
+## [0.3.3](https://github.com/Blobfolio/dactyl/releases/tag/v0.3.3) - 2022-03-28
+
+### Added
+
+* `dactyl::NoHash` (for `HashMap`, `HashSet`)
+* `dactyl::traits::BytesToSigned` (slice to signed integer parsing)
+
+
+
 ## [0.3.2](https://github.com/Blobfolio/dactyl/releases/tag/v0.3.2) - 2022-03-27
 
 ### Added